pyproject.toml
[project]
name = "wechat-mcp-server"
version = "0.1.0"
description = "WeChat MCP server and helper scripts for reading and replying to messages via macOS Accessibility APIs"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Banghao Chi", email = "banghao2@illinois.edu" }]
dependencies = [
    "pyobjc>=12.1",
    "pyobjc-framework-applicationservices>=12.1",
    "pillow>=10.0.0",
    "mcp[cli]>=1.0.0",
    "openai>=1.0.0",
]

[project.scripts]
wechat-mcp = "wechat_mcp.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wechat_mcp"]

[tool.hatch.build]
packages = ["src/wechat_mcp"]
